# docs/swagger.yaml
basePath: /
definitions:
  models.ErrorResponse:
    properties:
      message:
        type: string
    type: object
  models.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
      user:
        $ref: '#/definitions/models.User'
    type: object
  models.User:
    properties:
      guid:
        type: string
      ip:
        type: string
    type: object
  services.TokenResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Authorization API Server
  title: Auth API
  version: "1.0"
paths:
  /auth/login/:
    post:
      consumes:
      - application/json
      description: authorization + token gen
      operationId: user-login
      parameters:
      - description: user info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: generated tokens
          schema:
            $ref: '#/definitions/services.TokenResponse'
        "400":
          description: bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Login
      tags:
      - songs
  /auth/refresh/:
    post:
      consumes:
      - application/json
      description: refresh tokens
      operationId: refresh-tokens
      parameters:
      - description: user info + refresh token
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: generated tokens
          schema:
            $ref: '#/definitions/services.TokenResponse'
        "400":
          description: bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Refresh
      tags:
      - songs
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"